rename fitting test files
